Only set DialogStatus.Closed when Dialog is really, truly gone.
Due to part of the disappearing animation for Dialog being controlled
by a different component, Closed status was incorrectly set before
the dialog had faded from the display.

In the case of dynamically-allocated dialogs destroyed on Closed, this
would leave the display with unexpected content (an unfaded black
rectangle, partially alpha) until the next full redraw.