Remove javascript font family variables and bindings.
Replace them with C++ properties instead. This removes the (completely useless,
as locale cannot change) runtime binding evaluation all over the place, as well
as centralising the declaration of the font families.

Technically, this is BIC, but UIConstants.js was never supposed to be public API
anyway.

This provides 4-5 FPS gain when panning large lists of labels.